refactor & resolve conversations
