+++ src/db/methods/__init__.py
@@ -1,1 +1,2 @@
-from .methods import *
+from .users import *
+from .organizations import *
