+++ src/t/__init__.py
@@ -1,5 +1,5 @@
-from .types import *
+from . import types
 from .requests import RQ
 from .responses import RS
 
-__all__ = ["RQ", "RS"]
+__all__ = ["types", "RQ", "RS"]
